user and auth routes testing, added admin auth

+++ pi/master/__init__.py
@@ -38,5 +38,6 @@ import master.sockets
 # create database (if not exist) before first api request
 @app.before_first_request
 def create_tables():
-    db.drop_all()
-    db.create_all()
+    if app.config['TESTING'] == False:
+        db.drop_all()
+        db.create_all()
